combine BarrierTaskContext with BarrierTaskContextImpl.
